Add files via upload
This is a 2 line program that automates WhatsApp using the python library 'pywhatkit'. it just takes the phone number to whom you want to send the WhatsApp, the message you want to send, and hour and minutes in military timing. If you went to bed early there is no need to wake up at midnight for birthdays. you just need to install a library and 'pywhatkit ' and the two lines of codes.
NOTE: I have difficulty using this library in Jupyter Notebook. I would suggest that you use this library in some IDE like pycharm, IDLE, or sublime text

In this program, I am using 'pywhatkit' to automate WhatsApp. There is more application of this library.